Get rid of `app/javascript/core`
Have all flavors implement everything they need instead.
