+++ app/controllers/statuses_controller.rb
@@ -41,7 +41,7 @@ class StatusesController < ApplicationController
   end
 
   def embed
-    use_pack 'embed'
+    use_pack 'public'
     return not_found if @status.hidden? || @status.reblog?
 
     expires_in 180, public: true
