--- app/javascript/core/auth.js
@@ -1,3 +1,0 @@
-import 'packs/public-path';
-import './settings';
-import './two_factor_authentication';
