+++ app/javascript/flavours/glitch/packs/common.js
@@ -1,4 +1,5 @@
 import 'packs/public-path';
+import 'font-awesome/css/font-awesome.css';
 import Rails from '@rails/ujs';
 import 'flavours/glitch/styles/index.scss';
