+++ app/javascript/flavours/glitch/packs/inert.js
@@ -1,0 +1,4 @@
+/* Placeholder file to have `inert.scss` compiled by Webpack
+   This is used by the `wicg-inert` polyfill */
+
+import '@/styles/inert.scss';
