+++ app/javascript/flavours/glitch/packs/mailer.js
@@ -1,0 +1,3 @@
+import '@/styles/mailer.scss';
+
+require.context('@/icons');
