+++ app/javascript/flavours/glitch/theme.yml
@@ -1,26 +1,32 @@
 #  (REQUIRED) The location of the pack files.
 pack:
   admin:
-    - packs/admin.tsx
-    - packs/public.tsx
-  auth: packs/public.tsx
+    - admin.tsx
+    - public.tsx
+  auth:
+    - public.tsx
+    - two_factor_authentication.js
   common:
-    filename: packs/common.js
+    filename: common.js
     stylesheet: true
-  embed: packs/public.tsx
-  error: packs/error.js
+  error: error.js
   home:
-    filename: packs/home.js
+    filename: home.js
     preload:
       - flavours/glitch/async/compose
       - flavours/glitch/async/home_timeline
       - flavours/glitch/async/notifications
+  inert:
+    filename: inert.js
+    stylesheet: true
   mailer:
-  modal:
-  public: packs/public.tsx
-  settings: packs/public.tsx
-  sign_up: packs/sign_up.js
-  share: packs/share.jsx
+    filename: mailer.js
+    stylesheet: true
+  public: public.tsx
+  settings: public.tsx
+  sign_up: sign_up.js
+  share: share.jsx
+  remote_interaction_helper: remote_interaction_helper.ts
 
 #  (OPTIONAL) The directory which contains localization files for
 #  the flavour, relative to this directory. The contents of this
@@ -34,6 +40,12 @@ inherit_locales: vanilla
 #  (OPTIONAL) A file to use as the preview screenshot for the flavour,
 #  or an array thereof. These are the full path from `app/javascript/`.
 screenshot: flavours/glitch/images/glitch-preview.png
+
+#  (OPTIONAL) The directory which contains the pack files.
+#  Defaults to this directory (`app/javascript/flavour/[flavour]`),
+#  but in the case of the vanilla Mastodon flavour the pack files are
+#  somewhere else.
+pack_directory: app/javascript/flavours/glitch/packs
 
 #  (OPTIONAL) The directory which contains the pack files.
 #  Defaults to the theme directory (`app/javascript/themes/[theme]`),
