+++ app/javascript/flavours/vanilla/theme.yml
@@ -3,11 +3,12 @@ pack:
   admin:
     - admin.tsx
     - public.tsx
-  auth: public.tsx
+  auth:
+    - public.tsx
+    - two_factor_authentication.js
   common:
     filename: common.js
     stylesheet: true
-  embed: public.tsx
   error: error.js
   home:
     filename: application.js
@@ -15,12 +16,17 @@ pack:
       - features/compose
       - features/home_timeline
       - features/notifications
+  inert:
+    filename: inert.js
+    stylesheet: true
   mailer:
-  modal:
+    filename: mailer.js
+    stylesheet: true
   public: public.tsx
   settings: public.tsx
   sign_up: sign_up.js
   share: share.jsx
+  remote_interaction_helper: remote_interaction_helper.ts
 
 #  (OPTIONAL) The directory which contains localization files for
 #  the flavour, relative to this directory.
