+++ app/javascript/packs/common.js
@@ -1,2 +1,5 @@
 import './public-path';
+import 'font-awesome/css/font-awesome.css';
 import 'styles/application.scss';
+
+require.context('../images/', true);
