+++ app/views/layouts/error.html.haml
@@ -6,7 +6,6 @@
     %title= safe_join([yield(:page_title), Setting.default_settings['site_title']], ' - ')
     %meta{ content: 'width=device-width,initial-scale=1', name: 'viewport' }/
     = javascript_pack_tag 'common', crossorigin: 'anonymous'
-    = render partial: 'layouts/theme', object: @core || { pack: 'common' }
     = render partial: 'layouts/theme', object: @theme || { pack: 'error', flavour: 'glitch', common: { pack: 'common', flavour: 'glitch', skin: 'default' } }
   %body.error
     .dialog
