+++ app/views/remote_interaction_helper/index.html.haml
@@ -1,4 +1,4 @@
 - content_for :header_tags do
   %meta{ name: 'robots', content: 'noindex' }/
 
-  = javascript_pack_tag 'core/remote_interaction_helper', crossorigin: 'anonymous'
+  = javascript_pack_tag 'flavours/vanilla/remote_interaction_helper', crossorigin: 'anonymous'
